Add files via upload
A dataset of credit card transactions done by different number of customers has been taken and by applying Logistic Regression
model of machine learning, an analysis is done to check whether a credit card transaction is legit or fraud. It aims to track down
credit card transaction data, which is done by detecting anomalies in the transaction data and and by training the machine learning
model we make detections. There are so many transactions fraud activites performed by phishers and the main victims of these are
those poeple who don't have much knowledge in today's technologies. So this model will help to find the frauds and keep a track of
them.